final2

+++ ft_strlen.c
@@ -14,13 +14,14 @@
 
 size_t	ft_strlen(char const *s)
 {
-	size_t	len;
+	int	len;
 
 	len = 0;
 	while (s[len] != '\0')
 		len++;
 	return (len);
 }
+
 // TO FIND THE LENGHT OF THE SOURCE
 
 // int	ft_strlen(char const *s)
